Fixed bugs with class, known issue with linking project
