+++ .cproject
@@ -166,13 +166,9 @@
 								<option IS_BUILTIN_EMPTY="false" IS_VALUE_EMPTY="false" id="com.st.stm32cube.ide.mcu.gnu.managedbuild.tool.cpp.compiler.option.includepaths.1578006477" name="Include paths (-I)" superClass="com.st.stm32cube.ide.mcu.gnu.managedbuild.tool.cpp.compiler.option.includepaths" useByScannerDiscovery="false" valueType="includePath">
 									<listOptionValue builtIn="false" value="../Drivers/CMSIS/Device/ST/STM32H7xx/Include"/>
 									<listOptionValue builtIn="false" value="../Drivers/CMSIS/Include"/>
-									<listOptionValue builtIn="false" value="../Inc"/>
 									<listOptionValue builtIn="false" value="../Drivers/STM32H7xx_HAL_Driver/Inc"/>
 									<listOptionValue builtIn="false" value="../Drivers/STM32H7xx_HAL_Driver/Inc/Legacy"/>
-									<listOptionValue builtIn="false" value="&quot;${ProjDirPath}/Hyperlib/src&quot;"/>
-									<listOptionValue builtIn="false" value="&quot;${workspace_loc:/HyperLib/src}&quot;"/>
-									<listOptionValue builtIn="false" value="&quot;${workspace_loc:/HyperLib/Inc}&quot;"/>
-									<listOptionValue builtIn="false" value="&quot;${workspace_loc:/HyperLib}&quot;"/>
+									<listOptionValue builtIn="false" value="&quot;C:\Users\stefa\OneDrive - UPV\Documentos\Firmware-H8\ST-LIB\Inc&quot;"/>
 								</option>
 								<inputType id="com.st.stm32cube.ide.mcu.gnu.managedbuild.tool.cpp.compiler.input.cpp.1718039217" superClass="com.st.stm32cube.ide.mcu.gnu.managedbuild.tool.cpp.compiler.input.cpp"/>
 							</tool>
