--- Src/HALAL/Models/PinModel/Pin.hpp
@@ -1,174 +1,0 @@
-/*
- * Pin.hpp
- *
- *  Created on: 19 oct. 2022
- *      Author: stefa
- */
-
-#ifndef HALAL_MODELS_PIN_HPP_
-#define HALAL_MODELS_PIN_HPP_
-
-#include "stm32h7xx_hal.h"
-#include "../../../../Inc/C++Utilities/CppUtils.hpp"
-
-enum GPIO_Pin{
-	PIN_0 = ((uint16_t)0x0001),
-	PIN_1 = ((uint16_t)0x0002),
-	PIN_2 = ((uint16_t)0x0004),
-	PIN_3 = ((uint16_t)0x0008),
-	PIN_4 = ((uint16_t)0x0010),
-	PIN_5 = ((uint16_t)0x0020),
-	PIN_6 = ((uint16_t)0x0040),
-	PIN_7 = ((uint16_t)0x0080),
-	PIN_8 = ((uint16_t)0x0100),
-	PIN_9 = ((uint16_t)0x0200),
-	PIN_10 = ((uint16_t)0x0400),
-	PIN_11 = ((uint16_t)0x0800),
-	PIN_12 = ((uint16_t)0x1000),
-	PIN_13 = ((uint16_t)0x2000),
-	PIN_14 = ((uint16_t)0x4000),
-	PIN_15 = ((uint16_t)0x8000),
-	PIN_ALL = ((uint16_t)0xFFFF)
-};
-
-enum Operation_Mode{
-	NOT_USED,
-	INPUT,
-	OUTPUT,
-	ANALOG,
-	ALTERNATIVE,
-};
-
-class Pin{
-public:
-	GPIO_TypeDef * port;
-	GPIO_Pin pin;
-	Operation_Mode mode = Operation_Mode::NOT_USED;
-	static vector<reference_wrapper<Pin>> pinVector;
-
-	Pin();
-	Pin(GPIO_TypeDef* Port, GPIO_Pin Pin);
-	static void register_pin(Pin& pin, Operation_Mode mode);
-	static void unregister_pin(Pin& pin);
-	static void start();
-};
-
-Pin PE2(GPIOE,GPIO_Pin::PIN_2);
-Pin PE3(GPIOE,GPIO_Pin::PIN_3);
-Pin PE4(GPIOE,GPIO_Pin::PIN_4);
-Pin PE5(GPIOE,GPIO_Pin::PIN_5);
-Pin PE6(GPIOE,GPIO_Pin::PIN_6);
-Pin PC13(GPIOC,GPIO_Pin::PIN_13);
-Pin PC14(GPIOC,GPIO_Pin::PIN_14);
-Pin PC15(GPIOC,GPIO_Pin::PIN_15);
-Pin PF0(GPIOF,GPIO_Pin::PIN_0);
-Pin PF1(GPIOF,GPIO_Pin::PIN_1);
-Pin PF2(GPIOF,GPIO_Pin::PIN_2);
-Pin PF3(GPIOF,GPIO_Pin::PIN_3);
-Pin PF4(GPIOF,GPIO_Pin::PIN_4);
-Pin PF5(GPIOF,GPIO_Pin::PIN_5);
-Pin PF6(GPIOF,GPIO_Pin::PIN_6);
-Pin PF7(GPIOF,GPIO_Pin::PIN_7);
-Pin PF8(GPIOF,GPIO_Pin::PIN_8);
-Pin PF9(GPIOF,GPIO_Pin::PIN_9);
-Pin PF10(GPIOF,GPIO_Pin::PIN_10);
-Pin PH0(GPIOH,GPIO_Pin::PIN_0);
-Pin PH1(GPIOH,GPIO_Pin::PIN_1);
-Pin PC0(GPIOC,GPIO_Pin::PIN_0);
-Pin PC1(GPIOC,GPIO_Pin::PIN_1);
-Pin PC2(GPIOC,GPIO_Pin::PIN_2);
-Pin PC3(GPIOC,GPIO_Pin::PIN_3);
-Pin PA0(GPIOA,GPIO_Pin::PIN_0);
-Pin PA1(GPIOA,GPIO_Pin::PIN_1);
-Pin PA2(GPIOA,GPIO_Pin::PIN_2);
-Pin PA3(GPIOA,GPIO_Pin::PIN_3);
-Pin PA4(GPIOA,GPIO_Pin::PIN_4);
-Pin PA5(GPIOA,GPIO_Pin::PIN_5);
-Pin PA6(GPIOA,GPIO_Pin::PIN_6);
-Pin PA7(GPIOA,GPIO_Pin::PIN_7);
-Pin PC4(GPIOC,GPIO_Pin::PIN_4);
-Pin PC5(GPIOC,GPIO_Pin::PIN_5);
-Pin PB0(GPIOB,GPIO_Pin::PIN_0);
-Pin PB1(GPIOB,GPIO_Pin::PIN_1);
-Pin PB2(GPIOB,GPIO_Pin::PIN_2);
-Pin PF11(GPIOF,GPIO_Pin::PIN_11);
-Pin PF12(GPIOF,GPIO_Pin::PIN_12);
-Pin PF13(GPIOF,GPIO_Pin::PIN_13);
-Pin PF14(GPIOF,GPIO_Pin::PIN_14);
-Pin PF15(GPIOF,GPIO_Pin::PIN_15);
-Pin PG0(GPIOG,GPIO_Pin::PIN_0);
-Pin PG1(GPIOG,GPIO_Pin::PIN_1);
-Pin PE7(GPIOE,GPIO_Pin::PIN_7);
-Pin PE8(GPIOE,GPIO_Pin::PIN_8);
-Pin PE9(GPIOE,GPIO_Pin::PIN_9);
-Pin PE10(GPIOE,GPIO_Pin::PIN_10);
-Pin PE11(GPIOE,GPIO_Pin::PIN_11);
-Pin PE12(GPIOE,GPIO_Pin::PIN_12);
-Pin PE13(GPIOE,GPIO_Pin::PIN_13);
-Pin PE14(GPIOE,GPIO_Pin::PIN_14);
-Pin PE15(GPIOE,GPIO_Pin::PIN_15);
-Pin PB10(GPIOB,GPIO_Pin::PIN_10);
-Pin PB11(GPIOB,GPIO_Pin::PIN_11);
-Pin PB12(GPIOB,GPIO_Pin::PIN_12);
-Pin PB13(GPIOB,GPIO_Pin::PIN_13);
-Pin PB14(GPIOB,GPIO_Pin::PIN_14);
-Pin PB15(GPIOB,GPIO_Pin::PIN_15);
-Pin PD8(GPIOD,GPIO_Pin::PIN_8);
-Pin PD9(GPIOD,GPIO_Pin::PIN_9);
-Pin PD10(GPIOD,GPIO_Pin::PIN_10);
-Pin PD11(GPIOD,GPIO_Pin::PIN_11);
-Pin PD12(GPIOD,GPIO_Pin::PIN_12);
-Pin PD13(GPIOD,GPIO_Pin::PIN_13);
-Pin PD14(GPIOD,GPIO_Pin::PIN_14);
-Pin PD15(GPIOD,GPIO_Pin::PIN_15);
-Pin PG2(GPIOG,GPIO_Pin::PIN_2);
-Pin PG3(GPIOG,GPIO_Pin::PIN_3);
-Pin PG4(GPIOG,GPIO_Pin::PIN_4);
-Pin PG5(GPIOG,GPIO_Pin::PIN_5);
-Pin PG6(GPIOG,GPIO_Pin::PIN_6);
-Pin PG7(GPIOG,GPIO_Pin::PIN_7);
-Pin PG8(GPIOG,GPIO_Pin::PIN_8);
-Pin PC6(GPIOC,GPIO_Pin::PIN_6);
-Pin PC7(GPIOC,GPIO_Pin::PIN_7);
-Pin PC8(GPIOC,GPIO_Pin::PIN_8);
-Pin PC9(GPIOC,GPIO_Pin::PIN_9);
-Pin PA8(GPIOA,GPIO_Pin::PIN_8);
-Pin PA9(GPIOA,GPIO_Pin::PIN_9);
-Pin PA10(GPIOA,GPIO_Pin::PIN_10);
-Pin PA11(GPIOA,GPIO_Pin::PIN_11);
-Pin PA12(GPIOA,GPIO_Pin::PIN_12);
-Pin PA13(GPIOA,GPIO_Pin::PIN_13);
-Pin PA14(GPIOA,GPIO_Pin::PIN_14);
-Pin PA15(GPIOA,GPIO_Pin::PIN_15);
-Pin PC10(GPIOC,GPIO_Pin::PIN_10);
-Pin PC11(GPIOC,GPIO_Pin::PIN_11);
-Pin PC12(GPIOC,GPIO_Pin::PIN_12);
-Pin PD0(GPIOD,GPIO_Pin::PIN_0);
-Pin PD1(GPIOD,GPIO_Pin::PIN_1);
-Pin PD2(GPIOD,GPIO_Pin::PIN_2);
-Pin PD3(GPIOD,GPIO_Pin::PIN_3);
-Pin PD4(GPIOD,GPIO_Pin::PIN_4);
-Pin PD5(GPIOD,GPIO_Pin::PIN_5);
-Pin PD6(GPIOD,GPIO_Pin::PIN_6);
-Pin PD7(GPIOD,GPIO_Pin::PIN_7);
-Pin PG9(GPIOG,GPIO_Pin::PIN_9);
-Pin PG10(GPIOG,GPIO_Pin::PIN_10);
-Pin PG11(GPIOG,GPIO_Pin::PIN_11);
-Pin PG12(GPIOG,GPIO_Pin::PIN_12);
-Pin PG13(GPIOG,GPIO_Pin::PIN_13);
-Pin PG14(GPIOG,GPIO_Pin::PIN_14);
-Pin PG15(GPIOG,GPIO_Pin::PIN_15);
-Pin PB3(GPIOB,GPIO_Pin::PIN_3);
-Pin PB4(GPIOB,GPIO_Pin::PIN_4);
-Pin PB5(GPIOB,GPIO_Pin::PIN_5);
-Pin PB6(GPIOB,GPIO_Pin::PIN_6);
-Pin PB7(GPIOB,GPIO_Pin::PIN_7);
-Pin PB8(GPIOB,GPIO_Pin::PIN_8);
-Pin PB9(GPIOB,GPIO_Pin::PIN_9);
-Pin PE0(GPIOE,GPIO_Pin::PIN_0);
-Pin PE1(GPIOE,GPIO_Pin::PIN_1);
-
-
-
-
-#endif /* HALAL_MODELS_PIN_HPP_ */
